# Optimize Prometheus queries