Prometheus queries from the Telemetry service were taking seconds or 10s
of seconds.

Optimize these queries:
- Move all summary queries requiring a single data point off of Prometheus'
  QueryRange() endpoint, onto Query()
- Set `defaultVectorRange` to 30s, and also use it regardless of time
  window
Also add tests for grpc_server and telemetry server

Signed-off-by: Andrew Seigner <siggy@buoyant.io>

Fixes #260

# Results

Previously we had 12 variants of Prometheus query with p50's ranging from 4s to 8s. After these changes, all queries are < 3s, and all but 9 are < 300ms. p95 and p99 are also cut in half.

All results were generated using https://github.com/runconduit/conduit/tree/siggy/viz

## Average latency across all Prometheus queries

### Before

![sum_avg_baseline](https://user-images.githubusercontent.com/236915/35950577-111ff256-0c2c-11e8-9c42-560c54737ac0.png)

### After

![screen shot 2018-02-07 at 10 15 26 am](https://user-images.githubusercontent.com/236915/35950578-156c5a84-0c2c-11e8-9397-4cfd45d7dafb.png)

## Latency per Prometheus query

### Before

![avg_baseline_step](https://user-images.githubusercontent.com/236915/35950449-7cfeecee-0c2b-11e8-9c73-c7637a0ba6fd.png)

### After

![avg_query_step](https://user-images.githubusercontent.com/236915/35950451-80280edc-0c2b-11e8-9779-befaeac28d97.png)